BACKEND/modules/sastre/requirements.txt
# SASTRE AI Dependencies
anthropic>=0.18.0
fastapi>=0.109.0
uvicorn>=0.27.0
aiohttp>=3.9.0
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Auth
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.1.0

# Monitoring
prometheus-client>=0.19.0

# SDK
claude-agent-sdk>=0.1.18

# Utilities
tenacity>=8.2.0
structlog>=24.1.0
